requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pdfplumber==0.10.3
pandas==2.1.3
python-docx==1.1.0
python-multipart==0.0.6
openpyxl==3.1.2
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.0